feat: 사용 시간 저장

+++ src/entities/problem.go
@@ -25,7 +25,7 @@ type SaveSubmitResultDTO struct {
 	SubmitId   int
 	Result     string
 	UsedMemory int
-	UsedTime   int
+	UsedTime int64
 }
 
 type RunProblemRequest struct {
